+++ fastai/text/data.py
@@ -319,14 +319,14 @@ class TextList(ItemList):
     _processor = [TokenizeProcessor, NumericalizeProcessor]
     _is_lm = False
 
-    def __init__(self, items:Iterator, vocab:Vocab=None, pad_idx:int=1, **kwargs):
+    def __init__(self, items:Iterator, vocab:Vocab=None, pad_idx:int=1, sep=' ', **kwargs):
         super().__init__(items, **kwargs)
-        self.vocab,self.pad_idx = vocab,pad_idx
+        self.vocab,self.pad_idx,self.sep = vocab,pad_idx,sep
         self.copy_new += ['vocab', 'pad_idx']
 
     def get(self, i):
         o = super().get(i)
-        return Text(o, self.vocab.textify(o))
+        return Text(o, self.vocab.textify(o, self.sep))
 
     def label_for_lm(self, **kwargs):
         "A special labelling method for language models."
@@ -453,7 +453,7 @@ class SPProcessor(PreProcessor):
     def process_one(self, item, join=True):
         if join: text = _join_texts([item], self.mark_fields, self.include_bos, self.include_eos)[0]
         text = apply_rules(text, pre_rules=self.pre_rules, post_rules=self.post_rules)
-        return self._encode_batch([text])
+        return self._encode_batch([text])[0]
 
     def process(self, ds):
         ds.items = _join_texts(ds.items, self.mark_fields, self.include_bos, self.include_eos)